fix: correct PlayerController.Health return type (int->float), fix SendWebRequest return type, fix CD cache key
- PlayerController.Health stubbed as int but mod code uses float arithmetic
- UnityWebRequest.SendWebRequest() returns UnityWebRequestAsyncOperation in real Unity 2022.3
- CD pipeline stub cache key missing Assembly-CSharp_stubs.cs and Dread.csproj

Both stub return type mismatches cause MissingMethodException at runtime (CLR
treats return type as part of method signature).

+++ .github/scripts/Assembly-CSharp_stubs.cs
@@ -25,7 +25,7 @@ public class PlayerController : MonoBehaviour
     public float EnergyCurrent;
     public float EnergyStart;
     public float SprintSpeedMultiplier;
-    public int Health { get; }
+    public float Health { get; }
     public float stamina { get; }
 }
 public static class SemiFunc
